backend/api/requirements.txt
influxdb-client>=1.38.0
python-dotenv>=1.0.0
